html/cross-origin-opener-policy/blob-popup.https.html

<!doctype html>
<title>Cross-Origin-Opener-Policy and a blob URL popup</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/get-host-info.sub.js></script>
<script src="/common/utils.js"></script>
<script>
async_test(t => {
  window.test = t; // Make the test available globally so the blob URL can use it
  window.furtherPopup = null;

  const bc = new BroadcastChannel(token());
  bc.onmessage = t.step_func_done(({ data }) => {
    assert_equals(data.name, "");
    assert_false(data.opener);
    assert_true(furtherPopup.closed);
    assert_equals(furtherPopup.document.URL, "about:blank");
  });

  const blobContents = `<script>
const w = window.open("${get_host_info().HTTPS_REMOTE_ORIGIN}/html/cross-origin-opener-policy/resources/coop-coep.py?coop=x&coep=x&channel=${bc.name}", "${bc.name}");
window.opener.furtherPopup = w;

// Close the popup once the test is complete.
// The window proxy is closed hence use the broadcast channel to trigger the closure.
window.opener.test.add_cleanup(() => {
  bc.postMessage( "close" );
});
<\/script>`;
  const blob = new Blob([blobContents], { type: "text/html" });
  const blobURL = URL.createObjectURL(blob);
  const popup = window.open(blobURL);
  t.add_cleanup(() => popup.close());
  popup.onload = t.step_func(() => {
    assert_equals(popup.opener, window);
    assert_equals(popup.location.href, blobURL);
    assert_equals(popup.document.URL, blobURL);
    assert_equals(popup.origin, window.origin);
  });
});
</script>
