third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/blob-popup.https-expected.txt

This is a testharness.js-based test.
FAIL Cross-Origin-Opener-Policy and a blob URL popup assert_equals: expected 0 but got 36
Harness: the test ran to completion.
